bulding_agent_adk/pyproject.toml
[project]
name = "bulding-agent-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.25.1",
    "pandas>=3.0.1",
    "rich>=14.3.3",
    "tabulate>=0.9.0",
]
